+++ ext/StochasticSystemsBase.jl
@@ -2,13 +2,14 @@ module StochasticSystemsBase
 
 using DynamicalSystemsBase: DynamicalSystemsBase, SciMLBase, correct_state, CoupledODEs,
     CoupledSDEs, StateSpaceSets, isinplace, _delete, set_parameter!,
-    set_state!, dynamic_rule, isdeterministic, current_state,
+    set_parameters!, set_state!, dynamic_rule, isdeterministic, current_state,
     DynamicalSystemsBase, _set_parameter!, u_modified!,
     additional_details, referrenced_sciml_prob, DEFAULT_DIFFEQ,
-    SVector, SMatrix, current_parameters
+    SVector, SMatrix, current_parameters, initial_state, initial_time
 using SciMLBase: SDEProblem, AbstractSDEIntegrator, __init, SDEFunction, step!
-using StochasticDiffEq: SOSRA
+using StochasticDiffEq: SOSRA, setup_next_step!
 using LinearAlgebra
+import Random
 
 include("src/CoupledSDEs.jl")
 include("src/classification.jl")